# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
  /api/auth/magic/request:
    post:
      operationId: requestMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMagicLinkRequest"
      responses:
        "201":
          description: Created local magic-link token
  /api/auth/magic/consume:
    post:
      operationId: consumeMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumeMagicLinkRequest"
      responses:
        "200":
          description: Created session
...
  /api/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Workspace
...
  /api/channels/{channel_id}/messages:
    get:
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/channel_id"
        - name: after_seq
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Root channel messages
    post:
      operationId: createMessage
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "200":
          description: Existing message returned for an idempotent nonce replay
        "201":
          description: Created message
  /api/channels/{channel_id}/read:
    post:
      operationId: markChannelRead
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadRequest"
      responses:
        "200":
          description: Updated channel read receipt
...
  /api/messages/{message_id}/attachments:
    post:
      operationId: attachUpload
      parameters:
        - $ref: "#/components/parameters/message_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachUploadRequest"
      responses:
        "200":
          description: Attached upload to message
...
  /api/uploads:
    post:
      operationId: createUpload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [workspace_id, file]
              properties:
                workspace_id:
                  type: string
                file:
                  type: string
                  format: binary
                width:
                  type: integer
                  minimum: 0
                height:
                  type: integer
                  minimum: 0
                duration_ms:
                  type: integer
                  minimum: 0
      responses:
        "201":
          description: Created upload
  /api/uploads/{upload_id}:
    get:
      operationId: getUpload
      parameters:
        - name: upload_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload bytes
  /api/dms:
    get:
      operationId: listDirectConversations
      parameters:
        - name: workspace_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DM list
    post:
      operationId: createDirectConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDirectConversationRequest"
      responses:
        "201":
          description: Created DM
  /api/dms/{conversation_id}/messages:
    get:
      operationId: listDirectMessages
      parameters:
        - $ref: "#/components/parameters/conversation_id"
        - name: after_seq
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Direct messages
    post:
      operationId: createDirectMessage
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "200":
          description: Existing direct message returned for an idempotent nonce replay
        "201":
          description: Created direct message
  /api/dms/{conversation_id}/read:
    post:
      operationId: markDirectRead
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadRequest"
      responses:
        "200":
          description: Updated direct-conversation read receipt
...
  /api/hooks/slash/{channel_id}:
    post:
      operationId: slashCommand
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/SlashCommandRequest"
      responses:
        "201":
          description: Slash-command callback response
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    channel_id:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
    message_id:
      name: message_id
      in: path
      required: true
      schema:
        type: string
    conversation_id:
      name: conversation_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    RequestMagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
    ConsumeMagicLinkRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    CreateMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
        body_format:
          type: string
          enum: [markdown]
          default: markdown
        quoted_message_id:
          type: string
          description: |
            Optional ID of a message to quote. The quoted message must live in
            the same channel, DM conversation, or thread as the new message;
            cross-context quoting is rejected with 400. The server captures a
            soft snapshot of the quoted body at send time.
        nonce:
          type: string
          maxLength: 128
          description: |
            Optional client idempotency key for retry-safe message creation.
            Reusing the same nonce with the same body and quote returns the
            existing message with HTTP 200 instead of creating a duplicate.
    MarkReadRequest:
      type: object
      required: [seq]
      properties:
        seq:
          type: integer
          format: int64
          minimum: 0
          description: Last visible channel or DM sequence to mark as read.
...
    AttachUploadRequest:
      type: object
      required: [upload_id]
      properties:
        upload_id:
          type: string
    CreateDirectConversationRequest:
      type: object
      required: [workspace_id, member_ids]
      properties:
        workspace_id:
          type: string
        member_ids:
          type: array
          items:
            type: string
...
    SlashCommandRequest:
      type: object
      properties:
        command:
          type: string
        text:
          type: string
        user_name:
          type: string